refactor: change work with common module
