+++ src/common/formatString.js
@@ -1,0 +1,10 @@
+function formatString(string, parameters) {
+  let formattedString = string;
+  parameters.forEach((value, index) => {
+    formattedString = formattedString.replace(`{${index}}`, value);
+  });
+
+  return formattedString;
+}
+
+export { formatString };
